[ENG-1025] When decoding DocKey, return Corruption status instead of Check Failure
Summary:
ConsumePrimitiveValuesFromKey will return a bad status instead of DCHECK.

DecodeFrom will no longer attempt to Decode SeekKeys that are not correct encoding of good SubDocKeys. This is because it is not clearly defined what Seek keys may be like (May end in 0xFF? At DocKey or SubDocKey?) Decoding Seek Keys is only happening for verbose logging purposes from the BestEffortDocDBKeyToStr function. In case it's a seek key, whose decoding fails, the raw bytes will be returned, which is fine.

Test Plan: Existing tests

Reviewers: bharat, mikhail

Reviewed By: bharat, mikhail

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1568