Remove client_samples-test and client_symbol-test
Summary:
Remove client_samples-test and client_symbol-test and all special handling for them in the test
framework. The reason is that what's left of Kudu's C++ client is an internal client for us and we
have enough tests testing it already. The primary purpose of client_{samples,symbol}-test was to
test that the Kudu client code was self-sufficient and would compile without the server-side code,
which is not a priority for us.

Test Plan: Jenkins

Reviewers: pritam.damania, kannan, karthik, mihnea

Reviewed By: karthik

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D1564