Delete successful test output on CentOS
Summary:
Delete successful test output on CentOS. This is something we've been doing on Mac OS X for a while,
but it will also help alleviate the network bandwidth problem on the Jenkins master until we move
it to a bigger instance. Planning to move Jenkins master to c4.2xlarge from c3.xlarge.

Test Plan: Jenkins

Reviewers: bogdan, bharat, pritam.damania, kannan

Reviewed By: kannan

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1609