ENG-249: Update commit id from StartElection request and adapt test to run with leader failure check enabled
Summary:
There were 2 issues:
1) It is possible that new leader has all commited entries, but does not know that fact.
In this case we just try to update commit id from provided StartElection request.

2) Quite frequently it is not enough, because it could really mis some commited entries, so it could not become leader.
In this case leader failure check should handle situation.
It was disabled for this test, because it is quite straight forward.
I enabled this check, and adapted test to run with this check enabled.

ENG-249: Fix linter errors

Test Plan:
Run:
 ybd --cxx-test client_failover-itest --gtest_filter ClientFailoverITest.TestDeleteLeaderWhileScanning -n 100
It should not fail.

Reviewers: robert, bogdan, mikhail, bharat

Reviewed By: mikhail, bharat

Subscribers: kannan, bharat, jenkins-bot, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D1542