Set default softwareVersion to softwarVersions[0]
Summary:
We need to set default software version value once we receive a response from API server.
If response not present on mount we do logic in componentWillReceiveProps else we do it in componentWillMount()

Test Plan:
Go to create universe.
Choose the software version.
The correct software version should be used in user intent.

Reviewers: karthik, bharat, ram

Reviewed By: bharat, ram

Subscribers: yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D1624