Fix Select error <span/> cannot be child of <select/> add check for inconsistent universe with no regions to map regionUUID to
Summary:
When universe has region uuid, the region for which no longer exists, it will not be able to find the corresponding data object.

Test Plan:
Go to Create Universe Form, no browser console errors on changing the value of select.
Map should render even if the regions with given regionuuid does not exist.
{F3066}

Reviewers: karthik, ram

Reviewed By: ram

Subscribers: yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D1650