ENG-1179 Header highlight data disappears
Summary:
This is because a loading flag check was being done, the loading has since been changed from a flag to an object.
Also a loading indicator is not really needed to TopNavBar.

Test Plan:
Go to any page on Dashboard.
You should be able to see Highlighted Stats Panel

Reviewers: karthik, ram

Reviewed By: ram

Subscribers: shawn, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D1672