ENG-1216: Add error tracking and tracing info for some client wait API's.
Summary:
Noticed in one of the UI based expand universe that yugaware layer was stuck in `client.waitForLoadBalance()` call for >30min. So adding some logging and error info to that API. Noticed the same missing in waitForSever.

NOTE: Not de-duping the code for now, as it would just end up adding too many if/else in any common api...

Test Plan:
mvn test for yb-client, will run create/expand in UI now.

Reviewers: bogdan

Reviewed By: bogdan

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D1713