[ENG-1280] Support system.local table natively.
Summary:
The system.local table is implemented differently from the other system tables. We keep the
metadata information for the table in the master, but don't have any virtual table implementation
for it in the master. The reason for this is that it is much simpler to the return the data from the
CQL proxy itself for now. Later on, we could move this table to the master as we might need
additional information.

Test Plan: cqlsh, jenkins, java load test.

Reviewers: oleg, bogdan, robert

Reviewed By: robert

Subscribers: jenkins-bot, bharat, mikhail, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D1819