[ENG-1303][ENG-1331] Perf improvement and cleanup + refactor of Doc Rowwise Iterator
Summary:
Instead of working with rocksdb iterators directly, we'll use GetSubDocument from doc_rowwise_iterator. This effort already started in D1551, but also introduced perf regression. In this diff, we address it more comprehensively, and address perf issues.

In D1551 we were calling GetSubDocument for each column, which individually checks the row header each time (duplicate work). Now we have GetSubDocument for the whole row, that takes a projection, and avoids the duplicate work. A lot of machinery in doc_rowwise_iterator are no longer necessary because GetSubDocument handles them.

For selecting complete rows with n simple columns, We were previously seeking ~2n times. (Back and forth between top of row and each individual column). Now we should be seeking only once if necessary, and most others should be simple Next.

For rows with a lot of columns, if selecting a projection of n non-contiguous columns, we should be seeking n times to each column instead of 2n times that was happening previously.

@pritam.damania Will be primary reviewer for doc_rowwise_iterator, but I'm hoping others will be able to review high level. Also @robert had a good way of testing the number of seeks, I'll chat with him and test the performance together.

Test Plan: existing tests

Reviewers: mihnea, robert, pritam.damania, mikhail

Reviewed By: mikhail

Subscribers: amitanand, jenkins-bot, kannan, pritam.damania, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1839