Fix bug in system tablets initialization.
Summary:
The abstract tablets for each of the system tablet was initialized on startup only if the
appropriate tables weren't already present. The problem with this was that these tablets were null
when the cluster restarted and these tables were already present.

To fix, this I've modified the code to ensure that these tablets are always initialized.

Test Plan: unit test, jenkins

Reviewers: bogdan, robert

Reviewed By: robert

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D1846