ENG-1310 UniverseDetail popups causes page resize causing react-measure to fire
Summary:
React-bootstrap applies inline overflow: hidden class to Modal .
This is all well and good but it causes a page resize which in our case causes React-Measure to fire which sets of a chain of metrics querying delaying the entire page load.

Test Plan:
Go to Universe Detail page.
Click on any of the popups in the right while inside the metrics tab.
The popup should open and close smoothly

Reviewers: karthik, jeff, shawn, ram

Reviewed By: ram

Subscribers: jenkins-bot, shawn, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D1816