Remove unused imports/ Change React-Dom to latest version that suppresses warnings
Summary:
With React-Dom@15.5.0 we are getting warnings React.createClass is deprecated.
We are not using react.createClass in our code-base but some of the dependancies react-bootstrap, react-select use react.createClass and have PR's regarding the same have not been yet been resolved.
https://github.com/JedWatson/react-select/issues/1661
Hence keeping to React, React-Dom versions 15.4.0 until the PR's for React-Bootstrap and React-Select are resolved.

Test Plan:
Go to browser console.
You should not see any console warnings.

Reviewers: karthik, jeff, ram

Reviewed By: ram

Subscribers: jenkins-bot, ui, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D1890