ENG-1372: Run AWSInitializer on AppInit
Summary:
Due to changes from [[ https://yugabyte.atlassian.net/browse/ENG-898 | ENG-898 ]], there's an issue where users with an old version of the InstanceType db that hasn't been migrated to the new version will have a broken experience. Ultimately this is because instanceTypeDetails.volumeDetailsList will be empty. We need to run AWSInitializer to populate it (and change AWSInitializer to update instanceTypeDetails.volumeDetailsList, even if instanceTypeDetails is not null). This is done at AppInit so that the user doesn't have to explicitly send a REST request.

One resultant pain point is that logging in after starting the server will be slow. This is another reason we should rework the pricing details flow and probably put it into devops.

Test Plan:
sbt test; manually reset db to an old version, restart server and check that instance details are properly populated

Instance Type Details show up properly:
{F3557}

Reviewers: ram, bharat, karthik

Reviewed By: karthik

Subscribers: karthik, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D1934