ENG-1123: Group call responses
Summary:
Currently to send some data to a connection we place a task on the reactor.
And when this task is executed we add bytes to the send queue.
This could be optimized for the scenario when we receive additional requests
while waiting for our task to be executed in a reactor thread.
This diff addresses this for call responses.
The same behavior for outbound calls will be added later,
because this diff is already big enough.
Also it contains the following improvements:
1) Use same ev::io for read & write operations.
2) Try to write immediately, because usually we could do it w/o waiting.
3) Use RefCntBuffer for the send queue, so that the sender does not have to care about data ownership.
4) Send an outbound call using a single buffer for header & data.

Test Plan: Nothing specific, since those mechanics is involved in any test that requires RPC

Reviewers: robert, amitanand, mikhail

Reviewed By: mikhail

Subscribers: jenkins-bot, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D1823