[Eng-1340] Set timeout for Jedis client
Summary:
itest value-mismatch failures occur because the Jedis client seems to
timeout before the proxy layer does.

Set timeout for Jedis client to avoid this situation, and fix itests.

See Jira for possible alternatives.

We may not want to rely on this fix in the long term, as the clients may
want to have a different setting; and we should allow for a way to change
the proxy layer's timeout accordingly. YB Client already does this. But,
I don't think the redis protocol allows for such extra fields as timeout.
Will look in how they handle the issue and implement a long term fix.

Test Plan:
Do what itest does locally.

run workload locally, by starting a cluster; kicking off a cluster; and performing add node + restart-tservers a few times to repro this.

Reviewers: akashnil, karthik, kannan, bharat

Reviewed By: kannan, bharat

Subscribers: kannan, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D1952