ENG-1387: Prevent leader switch churn on node restart
Summary:
Added mechanic in Raft to reject Stepdown requests with hint for new leader if that
suggested leader is not caught up to the majority accepted OpId. That is because it would never be
able to become leader, should we force him to start an election.

This way, in the case of a node restart, the load balancer will still try to move leaders onto the
recently killed node, however, those requests will be rejected by the Raft leaders if the target
node is not yet fully caught up and the load balancer will retry in the next run.

Test Plan:
Started a cluster and workload, stopped a tserver for a bit then restarted it. Confirmed
that LB was sending Stepdown requests but they were being rejected according to new rule.

Reviewers: amitanand, bharat, kannan

Reviewed By: bharat, kannan

Subscribers: jenkins-bot, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D1968