[ENG-1390] Remove Original Write Requests from Replicate Messages
Summary:
We use WriteRequestPB for two purposes:
1) Communication from (YQL/Redis) proxy to Tserver.
2) Inside ReplicateMsgs when communicating from leader to follower.
The former contains a protobuf representation of the detailed request, the later contains on rocksdb-key-values, a post-image of executing the queries in the given transaction.
We unnecessarily kept the former request in the Replicate Msg which is not needed, only the post image of KVPairs need to be replicated.

Test Plan: Existing tests

Reviewers: robert, amitanand, kannan, mikhail

Reviewed By: mikhail

Subscribers: jenkins-bot, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D1948