[ENG-1438] Java integration test to support expand/shrink cluster
Summary:
This diff introduces a java integration test to expand and shrink a YB cluster with load
running. It performs the following:
1) Start load tester with 3 tservers.
2) Add 3 tservers and verify load spreads.
3) Remove original 3 tservers.
4) Verify load runs fine on new 3 tservers.

Apart from this, there are a few fixes that were needed to make this work:
1) When tablet bootstrap failed, the child memory tracker wasn't removed from the parent causing an
Assertion failure.
2) On tablet bootstrap failure, we weren't shutting down the tablet. This caused issues down the
line.
3) Blacklist for catalog_manager.cc wasn't updated correctly.

NOTE: This test is still flaky and probably needs a few fixes based on JIRAs that I've opened:
ENG-1454, ENG-1455, ENG-1447, ENG-1431. Although, I'd like to fix those issues in separate diffs so
that we iterate on the issues found rather than wait on this diff until all issues are fixed.

Test Plan: jenkins

Reviewers: karthik, hector, mikhail, bogdan, bharat

Reviewed By: bharat

Subscribers: jenkins-bot, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2021