ENG-1439: Fix YQLReadOperation to return static columns in continutation pages of a SELECT statement.
Summary: When selecting a static column and the result spans across pages, the static column can be omitted in the continuation pages. The value of the static column is always null. This happens becaue the DocYQLScanSpec built with the specific start key omits the static columns and we need a separate scan spec to fetch those static columns for the start key.

Test Plan: ybsql-static-column-test with test case for paging select.

Reviewers: bogdan, pritam.damania

Reviewed By: pritam.damania

Subscribers: kannan, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D1999