ENG-1445: Fix TestReplaceChangeConfigOperation: wait for leader election before bringing back the old leader
Summary:
The issue in ENG-1445 is that as we're trying to pause a leader and make it fall behind, and then
bring it back, we need to wait for the remaining servers to have already elected a leader.
Otherwise, with failure detector turned off, the old leader comes back and disrupts the election,
and we get stuck without a leader.

Also converting some comments to log messages to make test output more readable.

Test Plan:
- Jenkins
- `ybd release --cxx-test raft_consensus-itest --gtest_filter RaftConsensusITest.TestReplaceChangeConfigOperation -n 100 --clang`

Reviewers: robert, bogdan, bharat

Reviewed By: bharat

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2006