[ENG-1400] Lazily create iterator only if needed on write path
Summary:
SetPrimitive needs rocksdb iterators to read for init markers only if object init markers are required. In YQL, this is unnecessary. This change should avoid creating the iterator, which is where majority of cpu time is spent on the write path.

We want to rework docdb-iterators more generally, with laziness and other features, for now, it is hopefully a short-term perf fix.

Test Plan: manual testing using perf tool

Reviewers: kannan, mikhail

Reviewed By: mikhail

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D2063