ENG-1292 ENG-1110 ENG-1182 Fix KUDU/TabletPeerTest
Summary:
This fixes the following tests:
* KUDU/TabletPeerTest.TestDMSAnchorPreventsLogGC/0
* KUDU/TabletPeerTest.TestActiveTransactionPreventsLogGC/0
* KUDU/TabletPeerTest.TestMRSAnchorPreventsLogGC/0

The problem was that sometimes after we retrieve earliest_index and before we retrieve
last_log_opid, the log entry for leader election would be appended to the log (which comes from
StartPeer()). As a result, those two values would not match.

Now we wait for atleast one log entry to be appended before we move on.

Test Plan:
* ybd debug --cxx-test yb_tablet_peer-test --gtest_filter KUDU/TabletPeerTest.TestDMSAnchorPreventsLogGC/0 -n 100
* ybd debug --cxx-test yb_tablet_peer-test --gtest_filter KUDU/TabletPeerTest.TestMRSAnchorPreventsLogGC/0 -n 100
* ybd debug --cxx-test yb_tablet_peer-test --gtest_filter KUDU/TabletPeerTest.TestActiveTransactionPreventsLogGC/0 -n 100
* ybd debug --cxx-test yb_tablet_peer-test -n 500
* jenkins

Reviewers: timur, hector, mikhail

Reviewed By: mikhail

Subscribers: kannan, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2076