[ENG-1513] Fix comparable encoding for decimal zero
Summary:
Decimals are canonicalized to the form [sign] 10^x * 0.d1d2...dk where d1 and dk are non-zero. This method works for everything except zero. Two non-zero decimals with unequal exponents could always be compared with the exponent and sign without paying attention to the mantissa. But this comparison with exponent logic doesn't work for zero, because zero should ideally have exponent = -infinity (log(zero)). VarInt doesn't have a representation for -infinity. We had earlier forced zero to canonicalize as + 10^0 * 0.0. But the exponent zero is useless for comparison.

The only encoding for zero that will compare correctly with existing non-zero encodings is 0x80. But it is not possible to decode the size from that, because other decimals (like +10^-100000) starts with 0x80.

The only viable solution turned out to have two sign bits in the beginning of the encoding. Negatives (00), zero (10) and positives (11). Zero is specially encoded to 0x80 and decoded specially. The rest of the numbers are encoded and decoded just like before, but using two sign bits.

Also found that the decimal's own comparison operator was wrong for zero vs. 0.05, for the same reason (comparing exponents is not good for zeros, because the exponent is meaningless). Fixed that as well.

Test Plan: Added 0.05 as extra test value for encoding and comparison.

Reviewers: mikhail, oleg, hector

Reviewed By: hector

Subscribers: kannan

Differential Revision: https://phabricator.dev.yugabyte.com/D2086