ENG-1472: On LEARNER replica hitting `round not found` assert.
Summary:
For PRE_VOTER mode, the last received index was set but the committed index was not after a remote and then local bootstrap. And as such that round is not present in the pending_txns of a learner, and would have been ok if the committed index was also set after bootstrap...

Found the issue that the leader sending the logs and SSTables was taking a cut of the logs before the checkpointing the SSTables. So the last commit in the logs was older than the max commit in SST, in some cases. The fix is to take the cut of the SSTables before logs, any extra logs are processed as entries in the pending_txns_ on the PRE_VOTER side.

TODO: Add sanity check on sender for log and SST committed indexes. Fix a few more non-KUDU test failures in RBS tests...

Test Plan:
Enabled Java itest and confirmed it works - https://jenkins.dev.yugabyte.com/job/yugabyte-phabricator/442/testReport/org.yb.loadtester/TestClusterExpandShrink/

Testing others on jenkins.
```
#!/bin/bash
cd /Users/bharat/code/yugabyte/java/yb-loadtester; mvn clean
set -euo pipefail
for i in `seq 1 100`; do
   echo Running... $i
   mvn test -Dtest=org.yb.loadtester.TestClusterExpandShrink >& ~/misc/ces$i.txt
done
```

Reviewers: hector, mikhail

Reviewed By: mikhail

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2090