ENG-1123: Reading improvements
Summary:
Final diff in series of network layer improvements.
Redesign of receive. Previously for YB calls we were receiving in 2 phases.
First step is receiving size of packet and second step is receiving packet body.

It was changed to receive all available data, then parse it.
When we have a lot of calls, it is happens quite often that we receive data for several calls
with just a single system call.

Redesign of RPC classes was made, so we get rid of bunch of duplicated code.

@amitanand, please review redis part
@robert, please review CQL part

Test Plan: None specific.

Reviewers: robert, amitanand, mikhail

Reviewed By: mikhail

Subscribers: jenkins-bot, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2030