Print special-format log lines when every Java test starts and ends
Summary:
Print special-format log lines when every Java test starts and ends. Unlike we do in C++,
in Java we allow the test framework to run multiple test methods from the same class from a single
process (a single JVM), resulting in a long log. This is because in Java the program never crashes,
and exceptions are handled more cleanly. However, we need some way to split this long log so we can
look at individual tests' logs.

Test Plan: Jenkins

Reviewers: mihnea, bogdan, hector, bharat, timur

Reviewed By: bharat, timur

Subscribers: bharat, yql, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2111