ENG-1524 Add node list refresh event to CQL events sent to the client.
Summary:
When a node would be removed from the cluster, the CQL client would mark the node as being
down, but wouldn't remove it from its cluster metadata. With this diff, we ensure we send a periodic
full refresh of the CQL node list ensuring that any nodes that are not present in the system.peers
table are removed from the clients metadata.

Test Plan: unit test added.

Reviewers: mikhail, sergei

Reviewed By: sergei

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D2120