Refactoring of SignalRequest / SendNextRequest logic
Summary:
Making the logic of whether we're sending a request only if there are new operations in the queue or
unconditionally (as in case of heartbeats to peers) more understandable by replacing a bool with an
enum. Also updating / reformatting comments. Did this while reading the code.

Test Plan: Jenkins

Reviewers: bogdan, bharat, sergei

Reviewed By: sergei

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2126