ENG-1542: API for simulating network partitions in the mini cluster
Summary:
To test consistency in face of network partitions we need to be able to simulate network partitions in a mini-cluster.
This could be achieved by creating an API for enabling / disabling communication to / from a certain IP address.
So it is implementation of this API.

Test Plan: ybd --cxx-test rpc_rpc_stub-test --gtest_filter *.TestIncoherence

Reviewers: amitanand, pritam.damania, mikhail, bogdan

Reviewed By: mikhail

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2127