ENG-1481: Add a retry option to AddServer and RemoveServer test methods
Summary: Currently some tests fail when the leader is too slow and its committed term is not equal to its current term. When this happens. requests to change the config (add and remove servers) get rejected. This is only temporary, and we should keep retrying until the leader is ready.

Test Plan: Unit tests

Reviewers: timur, bharat

Reviewed By: bharat

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2147