ENG-1580 Introduce a flag to allow read from raft followers.
Summary:
This diff introduces a weaker consistency level, where we are allowed to read from raft
followers as long as the client explicitly requests this. For now, I've made changes on the tserver
layer to support such reads. Will follow this up with changes at the CQL proxy level to support this
for CQL clients.

Test Plan: jenkins

Reviewers: mikhail, hector, bogdan, bharat

Reviewed By: bharat

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2181