Resolve IP addresses for comparisons while populating system.local.
Summary:
The issue we were running into was that the master might be storing hostnames for TServer
registration, whereas the remote_endpont is retrieved from the RPC framework and is always an IP
address. As a result, we were performing string comparisons on both of these instead of resolving
the hostname to its IP addresses and checking if one matched.

Test Plan: test on AWS cluster.

Reviewers: karthik, oleg, robert, bogdan, sergei

Reviewed By: sergei

Subscribers: yql, jeff

Differential Revision: https://phabricator.dev.yugabyte.com/D2173