ENG-1213: Create new parse tree in case of reparse
Summary:
Reorg SQL engine to make Statement object read-only after parsing and semantic analysis. This will avoid self-deadlock on parse tree during statement execution when statement was locked in shared mode and needs to be relocked exclusively to reparse due to stale table metadata cache. To reparse, we will purge the table metadata cache and return UNPREPARED error to CQL client, upon which it will reprepare the statement.

In addition, we also purge the parse tree and create a new one now when a SQL statement needs to be re-analyzed because of stale table cache. This will be less error-prone than scrubbing the parse tree of analysis results and reparse.

Test Plan: TestInsert.testInsertIntoRecreatedTable Java test, ybsql-statement-test C++ test (updated).

Reviewers: amitanand

Reviewed By: amitanand

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D1696