ENG-1580 Repurpose CQL consistency levels to read from followers.
Summary:
This diff introduces support for reading from raft followers by specifying consistency
levels on the CQL client. LOCAL_ONE and QUORUM are repurposed to always read from the leader,
whereas ONE is used to read from followers. The reason we use LOCAL_ONE for strong consistency is
that this is the default for the CQL client and we don't want any client side changes.

Test Plan: unit tests.

Reviewers: hector, mikhail, bogdan, bharat, robert

Reviewed By: bharat, robert

Subscribers: robert, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2203