Support PostgreSQL 10 beta 1 step 1/2
The first half of the work to support PostgreSQL 10.

This just makes it can work with the version of PG. So partial index
paths are lost during Parallel hint processing.  Regression tests are
heavily modified but most of them just comes from the change of
original planner behavior introduced at the point of Pg10.