Fix AccessKey bug with key not being created
Summary: Tried to configure an OnPrem provider and found that the AccessKey wasn't being created. Root cause was a check on config.key.privateKeyContent to see if it's an object. But it's not an object, it's a String.

Test Plan: Try to configure an OnPrem provider through both JSON and Wizard flow. Verify AccessKey gets created in postgres db and /opt/yugaware/keys.

Reviewers: ram, pritam

Reviewed By: pritam

Subscribers: jenkins-bot, ui

Differential Revision: https://phabricator.dev.yugabyte.com/D2250