Fix heap-use-after-free in CQLServiceImpl::DeletePreparedStatementUnlocked().
Summary:
This issue is introduced by D2252 that added the function CQLServiceImpl::DeletePreparedStatementUnlocked(). The heap-use-after-free occurs because the function takes a "const shared_ptr&" of the CQL statement to delete from cache hash-map and LRU list. In the case when the shared_ptr is obtained from the LRU list from CQLServiceImpl::DeleteLruPreparedStatement(), the prepared_stmts_list_.erase() in DeletePreparedStatementUnlocked() will delete the very shared_ptr it is holding, causing the next "stmt->set_pos()" to refer to the deleted shared_ptr thus the issue.

To fix this issue, the "stmt" parameter is now a "const shared_ptr" and not a reference so that we hold a separate copy of the shared_ptr instead.

Test Plan: CQL ASAN Java tests.

Reviewers: mihnea

Reviewed By: mihnea

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D2292