Migrate Login/Register routing to BrowserHistory from React Context
Summary:
When API is slow to respond and we click submit multiple times, context.push appends the credentials to URL.
Migrating the routing to browserHistory should fix this -
https://stackoverflow.com/a/42007360
Either ways we should migrate to browserHistory since we are using browserHistory for all our routing elsewhere.
A good follow-up would be (depending on available cycles) migrating customer to {data: {}, status: {}, error: {}} format.
So that we can further control routing in componentWillReceiveProps

Test Plan: Login/Register flow should work as expected

Reviewers: jeff, ram

Reviewed By: ram

Subscribers: jenkins-bot, ui, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2283