ENG-1690 ENG-1728 Utilities to create appropriate tablet partitions for bulk load.
Summary:
This diff introduces a bunch of utilities that can be either used in a MR job or standalone
to create partitions for a MR job based on our partition scheme. Currently, it supports only a
handful of types that we need for customer-xyz. We can support additional types in the future.

Test Plan: jenkins

Reviewers: mihnea, mikhail, bogdan, karthik, sergei

Reviewed By: sergei

Subscribers: bharat, sergei, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2314