Bump 9.6 extension version to 1.1.1.
The 9.6 version of the extension was accidentally left at 1.0 which is used for the 9.5 version. The default is now set to 1.1.1 and an empty update script is included. Run:

alter extension pgaudit update to ‘1.1.1’

to update the extension version to 1.1.1, though there are no apparent problems with leaving it at 1.0.