ENG-1705 nodeDetailSet null causes universeDetail go to spinner
Summary:
This fixes the Bug which causes UI to go into Spinner when nodeDetailSet is null.
Fix is also contained in https://phabricator.dev.yugabyte.com/D2282
But that is a bigger diff and this fix is urgently needed.

Test Plan:
Go to universeDetail page.
It should not break when nodeDetailSet is null

Reviewers: karthik, ram

Reviewed By: ram

Subscribers: jenkins-bot, ui, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2365