ENG-1529: Promote a peer that was left in learner mode after it comes back
Summary: Currently if a peer is remote bootstrapped correctly, but it is restarted before it closes the connection with the leader, the leader will never issue a CHANGE_ROLE request to promote the peer from learner mode (PRE_VOTER or PRE_OBSERVER). To fix this, I have added code to the SendNextRequest method (the same method that sends a remote bootstrap request) to promote a peer when it detects that the peer doesn't need a remote bootstrap, but it's still in learner mode.

Test Plan: Added a new tests that simulates the situation described above

Reviewers: amitanand, bogdan, bharat

Reviewed By: bharat

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2212