ENG-1749: Instance type change during create universe should result in full move.
Summary:
When the total of node counts per-AZ was matching userIntent numNodes, we were missing check to ensure that there was in fact a change to at least one AZ node count (on the right panel). With the fix, if neither counts changed, then it falls back to full universe move, as expected for instance type change in this case.

Also did some cleanup to move all the pure expand/shrink logic into one API so that `updateUniverseDefinition` is cleaner. It works as expected for edit case as it is compared against persisted userIntent in the universe. Added tests (for create and edit cases).

Fix per AZ count in placement when configuring using user intent count. Fixed a couple of other issues found during testing.

Test Plan:
Added new tests. From UI went through the changes for nodes on both left and right side of modal and instance type changes.

Reviewers: karthik, pritam, jeff, ram

Reviewed By: ram

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2420