ENG-1690 Helper script for bulk load to copy SSTable files to production cluster.
Summary:
The script takes as input the tablet_id, tablet data directory, the replicas for the tablet
and the appropriate ssh key. It copies over the data to the appropriate nodes and then outputs the
replica with the corresponding directory where the data was copied. This would be used by tools to
import and cleanup that directory.

Test Plan: Jenkins: skip

Reviewers: bogdan, karthik, bharat

Reviewed By: bharat

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2472