[ENG-1730] Add Redis SREM and HDEL command
Summary:
We are adding SREM and HDEL command. No new protobufs are created, reusing RedisDelRequestPB with different KeyValue and types.

We check the type of any existing document, return error if trying to HDEL from a SET or some mismatch like that.

These commands are also subject to the flag emulate_redis_responses_. If the flag is set, we read the DB to figure out how many keys are actually being deleted and return it to the user. If not, then we return however many keys were requested to be deleted by the user.

Test Plan: Added tests in redisserver-test

Reviewers: karthik, amitanand

Reviewed By: amitanand

Subscribers: eng

Differential Revision: https://phabricator.dev.yugabyte.com/D2469