ENG-1774: Add flag for redis safe batching
Summary:
Currently we execute batches in parallel, that could be an issue if client send commands like this:
  set a 1
  get a
  set a 2
He expects that get returns 1 and that after execution of those commands a will have value of 2.

To resolve this issue we could track dependencies between commands and postpone execution of conflicting commands.

Test Plan: ybd --cxx-test redisserver_redisserver-test --gtest_filter *Safe*

Reviewers: amitanand, karthik, kannan, mikhail, akashnil

Reviewed By: akashnil

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2442