ENG-1690 By default use a high initial sequence number to always allow bulk load.
Summary:
Currently, the default for our initial sequence number is 0. This means if a customer
creates a cluster and wants to do a bulk load after the fact, this would never be possible. As a
result, we should always start with a high sequence number to allow enough room for bulk loads if it
is an after thought for some customer.

Test Plan: jenkins

Reviewers: mikhail, pritam.damania, sergei

Reviewed By: sergei

Subscribers: kannan, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2481