ENG-1809 : FATAL in existing tserver after adding new tserver while running CassandraTimeseries
Summary:
The man page of sem_gevalue(), used by consensus_peers, states the following.

```
If one or more processes or threads are blocked waiting to lock the semaphore with sem_wait(3), POSIX.1 permits two possibilities for the
value returned in sval: either 0 is returned; or a negative number whose absolute value is the count of the number of processes and
threads currently blocked in sem_wait(3). Linux adopts the former behavior. monospaced text
```

So on MacOS, it might return -1 when there are waiters (max is 1 waiter in our case). As such, the check can be modified so the value can be <= 0.

Itest and real clusters don't catch this as it is run against release binary.

Test Plan:
jenkins and steps in jira.

Reviewers: hector, robert

Reviewed By: robert

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2492